multer file upload feature, user schema: add profile features- Bio, profile picture; blog schema: add blog features- likes, comments; add relation between user and blogs; add controller and service logic for likes and comments; add some methods for file unlink and undefined route error handling; add uuid, and unique validator for unique users

+++ modules/Blogs/blogs.controller.js
@@ -1,10 +1,16 @@
-const blogSchema = require('../../db/db_config');
-const { addBlogService, getBlogByIdService, deleteBlogService } = require('./blogs.service');
+const { Blog, User } = require('../../db/db_config');
+const
+    { addBlogService, getBlogByIdService, deleteBlogService, addLikesService, addCommentsService }
+        = require('./blogs.service');
+
+const uuid = require('uuid').v4;
+const fs = require('fs');
+
 const mongoose = require('mongoose');
 
 module.exports = {
     getBlogs: (req, res) => {
-        blogSchema.find({}).then((blogs) => {
+        Blog.find({}).then((blogs) => {
             res.json(blogs);
         })
             .catch((err) => {
@@ -12,24 +18,40 @@ module.exports = {
                 res.status(500).json({ error: 'Error fetching Blogs' });
             });
     },
-    addBlog: (req, res) => {
-        const { title, content, creator } = req.body;
+    addBlog: async (req, res) => {
+        const { title, content, creator, tags, categories } = req.body;
 
-        if (!title || !content || !creator) {
+        if (!title || !content || !creator || !tags || !categories) {
             return res.status(400).json("Missing data fields! Could not create blog");
         }
 
-        addBlogService(title, content, creator, (err, result) => {
-            if (err) {
-                return res.status(500).send({ error: err });
+        try {
+            const user = await User.findById(creator);
+            if (!user) {
+                return res.status(404).json("User not found");
             }
-            else {
-                return res.status(201).send(result);
-            }
-        })
 
+            addBlogService(title, content, creator, tags, categories, (err, result) => {
+                if (err) {
+                    return res.status(500).send({ error: err });
+                } else {
+                    user.blogs.push(result._id);
+                    user.save().
+                        then(() => {
+                            res.status(201).send(result);
+                        })
+                        .catch((err) => {
+                            res.status(500).send({ error: 'Save Error' })
+                        });
+                }
+            });
+        }
+        catch (err) {
+            return res.status(500).send({ error: err.message });
+        }
 
     },
+
     getBlogById: (req, res) => {
         const blogId = req.params.blogId;
 
@@ -44,9 +66,38 @@ module.exports = {
 
     },
 
-    deleteBlog: (req, res) => {
+    deleteBlog: async (req, res, next) => {
         const blogId = new mongoose.Types.ObjectId(req.params.blogId);
         // const blogId=req.params.blogId;
+
+        let blog;
+        try {
+            blog = await Place.findById(blogId).populate('creator');
+        } catch (err) {
+            // const error = new HttpError(
+            //     'Something went wrong, could not delete place.',
+            //     500
+            // );
+            // return next(error);
+            return res.status(400).send({ error: err });
+        }
+
+        if (!blog) {
+            // const error = new HttpError("Could not find a place with that id.", 404);
+            // return next(error);
+            return res.status(400).send({ error: err });
+        }
+    
+        if(blog.creator.id !== req.blogId){
+            // const error = new HttpError( 'You do not have access to this API feature (deleting).', 401 )
+            // return next(error);
+            return res.status(400).send({ error: 'You do not have access to this API feature (deleting)' });
+        }
+
+        const imagePath = blog.image;
+        fs.unlink(imagePath, err => {
+            console.log(err);
+        });
 
         deleteBlogService(blogId, (err, result) => {
             if (err) {
@@ -56,5 +107,50 @@ module.exports = {
                 return res.status(200).json({ message: 'Successfully Deleted' });
             }
         })
+    },
+
+    addLikes: async (req, res) => {
+        try {
+            const blog = await Blog.findById(req.params.id);
+            if (!blog) {
+                return res.status(404).json({ error: 'Blog not found' });
+            }
+
+            addLikesService(blog, (err, result) => {
+                if (err) {
+                    return res.status(304).send({ error: err });
+                }
+                else {
+                    return res.status(200).json({ likes: result.likes });
+                }
+            });
+        } catch (error) {
+            return res.status(500).json({ error: error.message });
+        }
+    },
+
+    addComments: async (req, res) => {
+        try {
+            const blog = await Blog.findById(req.params.id);
+            if (!blog) {
+                return res.status(404).json({ error: 'Blog not found' });
+            }
+            const { name, comment } = req.body;
+            if (!name || !comment) {
+                return res.status(400).json({ error: 'Missing data fields' });
+            }
+
+            addCommentsService(blog, name, comment, (err, result) => {
+                if (err) {
+                    return res.status(304).send({ error: err });
+                }
+                else {
+                    return res.status(200).json(result);
+                }
+            });
+
+        } catch (error) {
+            return res.status(500).json({ error: error.message });
+        }
     }
 }
